requirements.txt
aiofiles==24.1.0
aiogram==3.22.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
aiosqlite==0.21.0
annotated-types==0.7.0
asyncpg==0.30.0
attrs==25.4.0
certifi==2025.10.5
ciso8601==2.3.3
frozenlist==1.8.0
idna==3.11
iso8601==2.1.0
magic-filter==1.0.12
multidict==6.7.0
orjson==3.11.4
propcache==0.4.1
pydantic==2.11.10
pydantic-settings==2.11.0
pydantic_core==2.33.2
pypika-tortoise==0.6.2
python-dotenv==1.2.1
pytz==2025.2
tortoise-orm==0.25.1
typing-inspection==0.4.2
typing_extensions==4.15.0
yarl==1.22.0
